pyproject.toml
[build-system]
requires = ["maturin>=0.12,<0.13"]
build-backend = "maturin"

[project]
name = "quanshu"
requires-python = ">=3.7"
classifiers = [
  "Programming Language :: Rust",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]

dependencies = ["uvloop == 0.16", "click == 8", "pyyaml"]

[project.optional-dependencies]
test = ["pytest > 5.0.0", "pytest-cov[all]"]

[project.urls]
homepage = "example.com"
documentation = "readthedocs.org"
repository = "github.com"
changelog = "github.com/me/spam/blob/master/CHANGELOG.md"

[project.scripts]
quanshu = "quanshu:main"
